requirements.txt
gradio>=4.0.0
opencv-python-headless
numpy
Pillow
moviepy
